Converted to a shared library
